Oh dear...

+++ AppSource/CUDALaplacePropagation.h
@@ -1,5 +1,4 @@
 #pragma once
-
 #include <vector>
 
 #include "Entity.h"
@@ -9,9 +8,9 @@ namespace CUDAHelpers
 	struct ComputingData
 	{
 		std::vector<float>& board;
-		int x_axis_bound;
-		int y_axis_bound;
-		int entity_radius;
+		uint32_t x_axis_bound;
+		uint32_t y_axis_bound;
+		uint16_t entity_radius;
 		Entity::EntityContainer swarm;
 	};
 
@@ -23,9 +22,9 @@ namespace CUDAHelpers
 		static void laplace(ComputingData data, Device device);
 
 	private:
-		static auto laplace_cpu(std::vector<float>& vec, int x_axis_bound, int y_axis_bound,
+		static auto laplace_cpu(std::vector<float>& vec, uint32_t x_axis_bound, uint32_t y_axis_bound,
 		                        Entity::EntityContainer swarm) -> void;
 
-		static auto laplace_gpu(std::vector<float>& vec, int x_axis_bound, int y_axis_bound) -> void;
+		static auto laplace_gpu(std::vector<float>& vec, uint32_t x_axis_bound, uint32_t y_axis_bound) -> void;
 	};
 }
